requirements.txt
fastapi==0.109.0
Pillow==9.5.0
requests==2.28.2
transformers==4.38.1
uvicorn==0.25.0
torch==2.2.0
